Read sensor values as signed integers
Temperature sensor readings are signed, which is hinted by the blank
value for them (oxd8, 216 as unsigned and -40 as signed). This is
confirmed by direct measurements using the a freezer, T_Sensor, and C8H
board.

Here we read fan sensors as signed too, because with their typical
values and 2-byte width, I can't tell a difference between signed and
unsigned, as I don't have a high speed chipset fan.